Created EventRegistry.java and PlayerJoinEventHandler.java
